rename: manageMember와 member 패키지 통합
