+++ src/main/java/com/example/scrap/converter/ManageMemberConverter.java
@@ -1,13 +1,13 @@
 package com.example.scrap.converter;
 
 import com.example.scrap.jwt.dto.Token;
-import com.example.scrap.web.manageMember.dto.ManageMemberResponse;
+import com.example.scrap.web.member.dto.MemberResponse;
 
 public class ManageMemberConverter {
 
-    public static ManageMemberResponse.ValidateTokenDTO toValidateTokenDTO(Token token){
+    public static MemberResponse.ValidateTokenDTO toValidateTokenDTO(Token token){
 
-        return ManageMemberResponse.ValidateTokenDTO.builder()
+        return MemberResponse.ValidateTokenDTO.builder()
                 .accessToken(token.getAccessToken())
                 .refreshToken(token.getRefreshToken())
                 .build();
