+++ src/main/java/com/example/scrap/web/member/MemberCommandServiceImpl.java
@@ -6,7 +6,9 @@ import com.example.scrap.entity.MemberLog;
 import com.example.scrap.entity.enums.SnsType;
 import com.example.scrap.web.category.CategoryRepository;
 import com.example.scrap.web.category.ICategoryCommandService;
+import com.example.scrap.web.member.dto.MemberDTO;
 import com.example.scrap.web.oauth.dto.NaverResponse;
+import com.example.scrap.web.scrap.IScrapCommandService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,12 +19,12 @@ import org.springframework.transaction.annotation.Transactional;
 public class MemberCommandServiceImpl implements IMemberCommandService {
 
     private final MemberRepository memberRepository;
+    private final IMemberQueryService memberQueryService;
     private final ICategoryCommandService categoryCommandService;
+    private final IScrapCommandService scrapCommandService;
 
     /**
      * 네이버 회원가입
-     * @param profileInfo
-     * @return
      */
     public Member signup(NaverResponse.ProfileInfo.Response profileInfo){
         MemberLog memberLog = new MemberLog();
@@ -35,9 +37,29 @@ public class MemberCommandServiceImpl implements IMemberCommandService {
     }
 
     /**
+     * 로그아웃
+     */
+    @Transactional
+    public void logout(MemberDTO memberDTO){
+        Member member = memberQueryService.findMember(memberDTO);
+
+        member.logout();
+    }
+
+    /**
      * 회원 탈퇴
      */
-    public void signOut(Member member){
+    @Transactional
+    public void signOut(MemberDTO memberDTO){
+        Member member = memberQueryService.findMember(memberDTO);
+
+        // 스크랩 전체 삭제
+        scrapCommandService.deleteAllScrap(memberDTO);
+
+        // 카테고리 전체 삭제
+        categoryCommandService.deleteAllCategory(memberDTO);
+
+        // 회원 삭제
         memberRepository.delete(member);
     }
 }
